fixed delete button

+++ client/src/components/Student.js
@@ -7,13 +7,16 @@ const Student = (props) => {
     const deleteStudent = (id) => {
 
         const url = `/api/v1.0/students/${id}`
-        axios.delete(url)
-            .then(response => {
-                this.props.history.push('/students')
-            })
-            .catch(error => {
-                console.log(error)
-            })
+        const confirm = window.confirm('Are you sure to delete? ')
+        if (confirm) {
+            axios.delete(url)
+                .then(response => {
+                    console.log(response)
+                })
+                .catch(error => {
+                    console.log(error)
+                })
+        }
 
     }
     return (
@@ -21,7 +24,6 @@ const Student = (props) => {
             <td>{props.student.name}</td>
             <td>{props.student.country}</td>
             <td>{props.student.age}</td>
-            <td>{props.student.bio}</td>
             <tr>
                 <NavLink to={`/students/${props.student._id}`} className="btn btn-primary">Detail</NavLink> {' '} {' '}
                 <NavLink to={`/edit/${props.student._id}`} className="btn btn-primary">Edit</NavLink> {' '} {' '}
